render CircularProgressIndicator while logging-in
